Upgrade to .NET8, divide existing code into clean architecture pattern.
