+++ jobs/Backend/Task/ExchangeRateUpdater.Console/ExchangeRateUpdater.Console.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <AssemblyName>ExchangeRateUpdater.Console</AssemblyName>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\ExchangeRateUpdater.Infrastucture\ExchangeRateUpdater.Infrastucture.csproj"/>
+    </ItemGroup>
+
+</Project>
