+++ jobs/Backend/Task/ExchangeRateUpdater.sln
@@ -3,7 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater", "ExchangeRateUpdater.csproj", "{7B2695D6-D24C-4460-A58E-A10F08550CE0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater.Domain", "ExchangeRateUpdater.Domain\ExchangeRateUpdater.Domain.csproj", "{E7D8A956-19EA-4DE6-981A-1D91E59BD3D0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater.Console", "ExchangeRateUpdater.Console\ExchangeRateUpdater.Console.csproj", "{8035EB61-EE7B-4817-A14D-83F1B32AC6C5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExchangeRateUpdater.Infrastucture", "ExchangeRateUpdater.Infrastucture\ExchangeRateUpdater.Infrastucture.csproj", "{1650DC68-CB0E-4E31-93CB-7B06C60915F2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +15,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7B2695D6-D24C-4460-A58E-A10F08550CE0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E7D8A956-19EA-4DE6-981A-1D91E59BD3D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7D8A956-19EA-4DE6-981A-1D91E59BD3D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7D8A956-19EA-4DE6-981A-1D91E59BD3D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7D8A956-19EA-4DE6-981A-1D91E59BD3D0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8035EB61-EE7B-4817-A14D-83F1B32AC6C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8035EB61-EE7B-4817-A14D-83F1B32AC6C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8035EB61-EE7B-4817-A14D-83F1B32AC6C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8035EB61-EE7B-4817-A14D-83F1B32AC6C5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1650DC68-CB0E-4E31-93CB-7B06C60915F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1650DC68-CB0E-4E31-93CB-7B06C60915F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1650DC68-CB0E-4E31-93CB-7B06C60915F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1650DC68-CB0E-4E31-93CB-7B06C60915F2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
